refactor($compile): 자동차 이름 원시값 포장 리팩터링
자동차 이름 원시값 포장 리팩터링
- CarName 클래스 생성
- Car 객체 생성시 문자열값을 CarName 객체 생성으로 리팩터링

자동차 이름 원시값 포장 리팩터링
